spliting componentItem out to its own python source file
